classifier/requirements.txt
torch==2.3.0
transformers[torch]==4.40.0
bentoml==1.3.0
python-dotenv==1.0.1